move pipe related functions to pipe.py
